% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceRoll.R
\name{constRoll}
\alias{constRoll}
\title{constant-value closure function}
\usage{
constRoll(adj)
}
\arguments{
\item{adj}{the constant the resulting functor should return}
}
\value{
a parameterless function that yields the adj constant
}
\description{
\code{constRoll} is a roll that always returns a predefined constant value.
This is useful in composing other roll functors.
}

