% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceRoll.R
\name{dcCheck}
\alias{dcCheck}
\title{Difficulty-Check roll}
\usage{
dcCheck(dc, mod = 0, checkRoll = NULL)
}
\arguments{
\item{dc}{a difficulty-check value to compare against}

\item{mod}{an optional modifier value (e.g. 10 for divine-guidance)}

\item{checkRoll}{an optional roll functor, defaults to 1D20}
}
\value{
Boolean success or failure roll
}
\description{
\code{dcCheck} makes a D20 roll against a difficulty-value.
}
\examples{
  saveVsPoisonCon13 <- dcCheck( 15, 1 )
  dwarfVsPoison <- dcCheck( 15, 3, withAdvantage() )

}

