% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceRoll.R
\name{effectiveHitDmg}
\alias{effectiveHitDmg}
\title{Calculates the effective average damage of your attacks
given a nominal damange, hitOdds, and critical-odds.
Be sure to add ability and misc (magic) bonuses to avgDmg,
e.g. effectiveHitDmg( mean(1:6)+3, hitodds( 17, 4, 3 ) )
for a proficiency 4 rogue with DEX mod 3 dmg and attack bonuses.}
\usage{
effectiveHitDmg(avgDmg, odds, critOdds = 0.05)
}
\description{
Calculates the effective average damage of your attacks
given a nominal damange, hitOdds, and critical-odds.
Be sure to add ability and misc (magic) bonuses to avgDmg,
e.g. effectiveHitDmg( mean(1:6)+3, hitodds( 17, 4, 3 ) )
for a proficiency 4 rogue with DEX mod 3 dmg and attack bonuses.
}

