% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceRoll.R
\name{hitOdds}
\alias{hitOdds}
\title{Calculates the probability you will hit a targetAC
given proficiency-bonus, ability-modifier, and misc adjustments
(e.g. magic-bonus)}
\usage{
hitOdds(targetAC, prof, ability, misc = 0)
}
\description{
Calculates the probability you will hit a targetAC
given proficiency-bonus, ability-modifier, and misc adjustments
(e.g. magic-bonus)
}

