% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChallengeRating.R
\name{listEncounters}
\alias{listEncounters}
\title{listEncounters matching a targetXp}
\usage{
listEncounters(targetXp)
}
\arguments{
\item{targetXp}{A scalar experience-point target for the encounter}
}
\value{
A dataframe of Baddies, CR, XP from 1 to 15 Baddies.
}
\description{
\code{listEncounters} Enumerates the combinations of foe counts and
 challenge-ratings that yield approximately the given  target XP value.
}
\examples{
   # A deadly encounter for 4 7th-level players
   listEncounters( 6800 )

   # A boss-and-minions battle of 4500(boss) and 2300(minions)
   # (This example won't compensate for $Boss$Baddies + $Minions$Baddies
   #  posing more difficulty than either $Baddies count alone, so be kind,
      pick rows that sum to 66-75\% targetXp, but still award targetXp)
   sapply( c(Boss=4500, Minions=2300),
           function(x) { listEncounters(x) },
           simplify=FALSE )
}
\seealso{
\code{\link{partyEncounter}}
}

