% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceRoll.R
\name{makeRoll}
\alias{makeRoll}
\title{makeRoll die-roll closure function}
\usage{
makeRoll(dieCt, dieType, adj = 0, ...)
}
\arguments{
\item{dieCt}{a list of counts of dice}

\item{dieType}{a list of dice sizes, e.g. 8 for a 1D8 die}

\item{adj}{a constant to also add to the sum of die-rolls}
}
\value{
a parameterless function that will make the roll when called
}
\description{
\code{makeRoll} creates a parameterless function that sums predefined
 sets of dice along with a constant adjustment.
}
\examples{
  threeD8    <- makeRoll( 3, 8 )
  \\code{#} To your D20, add an extra 1D4 and your 6 proficiency + ability bonus
  blessedAtk <- makeRoll( c(1,1), c(20,4), 6 )

}
\seealso{
\code{\link{multiRoll}}
}

