% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monsterCRExp}
\alias{monsterCRExp}
\title{Experience-points by Challenge-rating.}
\format{A data frame with 27 rows and 2 variables:
\describe{
  \item{CR}{challenge-rating}
  \item{XP}{experience-points}
  ...
}}
\source{
Monster Manual 5th Edition
}
\usage{
monsterCRExp
}
\description{
Representative XP worth of monsters by challenge-rating.
}
\keyword{datasets}

