% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monsterMultiCR}
\alias{monsterMultiCR}
\title{Experience-point multipliers for multiple-creature encounters.}
\format{A data frame with 15 rows and 2 variables:
\describe{
  \item{Baddies}{number of opponents}
  \item{Multiplier}{multiplicative factor of each opponent's XP}
  ...
}}
\source{
Dungeon Master's Guide 5th Edition
}
\usage{
monsterMultiCR
}
\description{
Difficulty increases super-linearly with the number of opponents,
at certain bands.
}
\keyword{datasets}

