% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceRoll.R
\name{multiRoll}
\alias{multiRoll}
\title{composite die-roll closure function}
\usage{
multiRoll(rollFuncList, evalFunc = sum)
}
\arguments{
\item{rollFuncList}{of die-roll functors to combine}

\item{evalFunc}{optional combining operation (default is sum)}
}
\value{
a parameterless function that will make the combined roll
}
\description{
\code{multiRoll} creates a parameterless functor that sums a list of
other rolls, e.g. composite effects from weapons and magic.
}
\examples{
  # Short-sword two-weapon fighting with for 3rd level rogue
  shortSwordTwFRoll <- multiRoll(  c(makeRoll(1,6,2), makeRoll(1,6,0)) )
  advantageAttack   <- multiRoll(  rep(

}
\seealso{
\code{\link{multiRoll}}
}

