% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChallengeRating.R
\name{partyEncounter}
\alias{partyEncounter}
\title{partyEncounter target XP calculator}
\usage{
partyEncounter(cts, lvls, difficulty = "Medium")
}
\arguments{
\item{cts}{A scalar or list of adventurer counts}

\item{lvls}{scalar or list (depending on cts) of adventurer levels}

\item{difficulty}{The encounter's inherent stress:
Easy, Medium, Hard, or Deadly (or All)}
}
\value{
The experience-points the combined monsters can sum to
        or a named-list for difficulty='All'.
}
\description{
\code{partyEncounter} computes the target XP value for an encounter with foes
 given a party of adventurers at different levels.
}
\examples{
   partyEncounter( c(3,1), c(5,7), 'Hard' )
   partyEncounter( 4, 3, 'All' )
}
\seealso{
\code{\link{listEncounters}}
}

