% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceRoll.R
\name{rollCheckElse}
\alias{rollCheckElse}
\title{Roll, validate, and re-roll}
\usage{
rollCheckElse(rollFunc, checkFunc, altFunc = NULL)
}
\arguments{
\item{rollFunc}{is the initial roll}

\item{checkFunc}{tests acceptance of the initial roll's value}

\item{altFunc}{is used if the checkFunc returns FALSE}
}
\value{
Either the rollFunc value or altFunc value
}
\description{
\code{rollCheckElse} examines the success of a first roll and either
uses the first value or makes an alternative roll.
}
\examples{
  # Use the D20 roll but re-roll on 1-2, and add 6 attack modifiers
  greatWeapFight <- rollCheckElse( makeRoll(1,20),
                                   function(r) { r > 2 } )
  myAttack <- multiRoll( c( greatWeapFight(), constRoll( 6 ) ) )

}

