% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xpThresholds}
\alias{xpThresholds}
\title{Target experience-point budgets per player.}
\format{A data frame with 80 rows and 3 variables:
\describe{
  \item{CharLevel}{character-level of player}
  \item{Difficulty}{encounter difficulty factor: Easy, Medium, Hard, Deadly}
  \item{Threshold}{budget-target this player can absorb, in XP}
  ...
}}
\source{
Dungeon Master's Guide 5th Edition
}
\usage{
xpThresholds
}
\description{
A dataset of character level experience-point targets for various difficulty regimes.
}
\keyword{datasets}

